Moved color converter functions to a new file. Split the image testcase to separate image io from color conversion.
