+++ mcdwt/image_io.py
@@ -1,5 +1,4 @@
 import cv2
-import numpy as np
 
 
 class InputFileException(Exception):
@@ -85,51 +84,3 @@ class ImageWritter:
 
         file_name = '{}{:03d}.png'.format(path, number)
         cv2.imwrite(file_name, image)
-
-
-class ImageColorConverter:
-    '''Convert PNG image color from YCbCr to RGB and vice versa.
-
-
-    '''
-
-    def __init__(self):
-        pass
-
-    def rgb2ycc(self, image):
-        '''Convert YCbCr to RGB
-
-            Parameters
-            ----------
-
-                image : [:,:,:].
-
-            Returns
-            -------
-
-                image : [:,:,:].
-
-                A image in YCbCr.
-        '''
-
-        imYCC = cv2.cvtColor(image, cv2.COLOR_BGR2YCR_CB)
-        return imYCC
-
-    def ycc2rgb(self, image):
-        '''Convert YCbCr to RGB
-
-            Parameters
-            ----------
-
-                image : [:,:,:].
-
-            Returns
-            -------
-
-                image : [:,:,:].
-
-                A image in RGB.
-        '''
-
-        imRGB = cv2.cvtColor(image, cv2.COLOR_YCR_CB2BGR)
-        return imRGB
